feat: Prepare team dashboard
