+++ src/main/java/org/owasp/dsomm/metricca/analyzer/controller/GrafanaTeamDashboardController.java
@@ -1,6 +1,7 @@
 package org.owasp.dsomm.metricca.analyzer.controller;
 
 import org.owasp.dsomm.metricca.analyzer.deserialization.ApplicationDirector;
+import org.owasp.dsomm.metricca.analyzer.grafana.TeamDashboard;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,11 +11,11 @@ import org.springframework.stereotype.Controller;
 public class GrafanaTeamDashboardController {
   private static final Logger logger = LoggerFactory.getLogger(GrafanaTeamDashboardController.class);
 
-//  @Autowired
-//  private ApplicationDirector applicationDirector;
-//
-//  @Autowired
-//  private TeamDashboard teamDashboard;
+  @Autowired
+  private ApplicationDirector applicationDirector;
+
+  @Autowired
+  private TeamDashboard teamDashboard;
 
 
 }
