+++ src/main/resources/application-dev.properties
@@ -11,4 +11,4 @@ metricCA.git.url=
 #metricCA.git.branch=feat/fixConstruction
 #metricCA.git.url=https://github.com/devsecopsmaturitymodel/metricAnalyzer.git
 
-metricCA.api.baseurl=http://192.168.178.23:8080
+metricCA.api.baseurl=http://192.168.2.114:8080
